requirements.txt
# Core dependencies
boto3>=1.28.0
pyspark>=3.4.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0

# Data validation
pydantic>=2.0.0
great-expectations>=0.17.0

# AWS Glue local development
aws-glue-libs>=4.0.0

# Utilities
python-dateutil>=2.8.2
faker>=19.0.0
click>=8.1.0
tqdm>=4.65.0
